Super fast server fix

+++ Utilities/Network.cpp
@@ -63,7 +63,7 @@ Parser::Messenger Network::inputNetwork(double time)
 	//jezeli czas trwania cyklu jest dluzszy od ustalonego, WSAPoll jest natychmiastowy
 	if (descrList.size() > 0)
 	{
-		result = WSAPoll(&this->descrList[0], this->descrList.size(), (int)time * 1000);
+		result = WSAPoll(&this->descrList[0], this->descrList.size(), (int)(time * 1000));
 	}
 	else
 	{
@@ -594,7 +594,7 @@ void Network::processMessage(Contact * client)
 			else
 				return;
 		}
-		if (client->bufferInput.size() >= client->msgExpectedLenght)
+		if ((int)client->bufferInput.size() >= client->msgExpectedLenght)
 		{
 			Logger::debug("Buffer is bigger or same size as message lenght.");
 			Parser::Event ev = Parser::decodeBytes(std::string(client->bufferInput, 0, client->msgExpectedLenght));
